MM-17915 Migrate edit channel to realm
